backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
celery>=5.2.7
redis>=4.5.4
pytesseract>=0.3.10
opencv-python>=4.7.0
pdf2image>=1.16.3
pillow>=9.5.0
transformers>=4.28.1
sentencepiece>=0.1.99
boto3>=1.26.115
sqlalchemy>=2.0.9
pydantic>=2.0.0
python-dotenv>=1.0.0 